requirements.txt
Pillow==5.0.0
pygame==1.9.3
PyOpenGL==3.1.0
